requirements.txt
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
jinja2>=3.1.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
PyYAML>=6.0